Add files via upload
Business Objectives
Your aim is to identify the target customers for the term deposits from the pool of the bank’s existing customers. You should also capture the key driving factors (or driver variables) behind the successful conversion of a customer, i.e., the customer opening a term deposit account with the bank. Using this information, the bank would optimize investment in its future marketing campaign.

Results Expected
Provide a step-by-step approach and details of the analysis performed to derive the insights in the form of a PowerPoint presentation.

Identify the missing data, duplicate data, outliers, spelling inconsistencies, and use the appropriate methods to deal with these issues.

Identify the possible derived columns and, if required, convert any specific column to a required format for ease of analysis.

Explain the results of univariate, segmented univariate, bivariate analysis, etc. in business terms. In addition, identify the insights that will help explain how a specific variable is playing a key role in converting the customers to open the term deposit account with the bank.

Include visualisations and also summarise the most important results in the presentation. You are free to choose the graphs that explain the numerical/categorical variables.

Derive the required values based on the EDA of the dataset and find out the optimized solution for the future marketing campaign.